+++ mindspore/lite/tools/converter/preprocess/image_preprocess.cc
@@ -116,7 +116,7 @@ int CenterCrop(cv::Mat *image, int width, int height) {
 }
 
 int PreProcess(const preprocess::DataPreProcessParam &data_pre_process_param, const std::string &input_name,
-               int image_index, mindspore::tensor::MSTensor *tensor) {
+               size_t image_index, mindspore::tensor::MSTensor *tensor) {
   if (tensor == nullptr) {
     MS_LOG(ERROR) << "tensor is nullptr.";
     return RET_NULL_PTR;
@@ -155,7 +155,7 @@ int PreProcess(const preprocess::DataPreProcessParam &data_pre_process_param, co
   return RET_OK;
 }
 
-int PreProcess(const DataPreProcessParam &data_pre_process_param, const std::string &input_name, int image_index,
+int PreProcess(const DataPreProcessParam &data_pre_process_param, const std::string &input_name, size_t image_index,
                void **data, size_t *size) {
   if (data == nullptr || size == nullptr) {
     MS_LOG(ERROR) << "data or size is nullptr.";
