+++ mindspore/lite/tools/converter/preprocess/image_preprocess.h
@@ -35,11 +35,11 @@ int Resize(cv::Mat *image, int width, int height, cv::InterpolationFlags resize_
 int CenterCrop(cv::Mat *image, int width, int height);
 
 // NOTE:`data` must be use delete[] to free buffer.
-int PreProcess(const DataPreProcessParam &data_pre_process_param, const std::string &input_name, int image_index,
+int PreProcess(const DataPreProcessParam &data_pre_process_param, const std::string &input_name, size_t image_index,
                void **data, size_t *size);
 
 int PreProcess(const preprocess::DataPreProcessParam &data_pre_process_param, const std::string &input_name,
-               int image_index, mindspore::tensor::MSTensor *tensor);
+               size_t image_index, mindspore::tensor::MSTensor *tensor);
 
 int ImagePreProcess(const ImagePreProcessParam &image_preprocess_param, cv::Mat *image, void **data, size_t *size);
